+++ app/Http/Controllers/Ajax/GroupController.php
@@ -5,15 +5,26 @@ namespace App\Http\Controllers\Ajax;
 use App\Models\Group;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Events\AddEventLogs;
+use App\Models\Role;
+use Illuminate\Support\Facades\Auth;
 
 class GroupController extends Controller
 {
     public function create(Request $request){
+        if(!Role::granted('ref_doc_add')){//вызываем event
+            $msg = 'Попытка создания новой группы контрагентов!';
+            event(new AddEventLogs('access',Auth::id(),$msg));
+            return 'NO';
+        }
         if($request->isMethod('post')){
             $input = $request->except('_token'); //параметр _token нам не нужен
             $model = new Group();
             $model->fill($input);
             if($model->save()) {
+                $msg = 'Группа контрагентов '. $input['name'] .' была успешно добавлена!';
+                //вызываем event
+                event(new AddEventLogs('info',Auth::id(),$msg));
                 $row = Group::all()->last();
                 $content = '<tr id="'.$row->id.'">
                                 <th><a href="/groups/view/'.$row->id.'">'.$row->name.'</a></th>
@@ -37,8 +48,17 @@ class GroupController extends Controller
             $input = $request->except('_token'); //параметр _token нам не нужен
             $id = $request->input('id');
             $model = Group::find($id);
+            if(!Role::granted('ref_doc_edit')){
+                $msg = 'Попытка редактирования группы контрагентов '. $model->name;
+                //вызываем event
+                event(new AddEventLogs('access',Auth::id(),$msg));
+                return 'NO';
+            }
             $model->fill($input);
             if($model->update()) {
+                $msg = 'Группа контрагентов '. $input['name'] .' была обновлена!';
+                //вызываем event
+                event(new AddEventLogs('info',Auth::id(),$msg));
                 $arr = array('id'=>$id,'name'=>$request->input('name'));
                 return json_encode($arr, JSON_UNESCAPED_UNICODE);
             }
@@ -52,7 +72,16 @@ class GroupController extends Controller
         if($request->isMethod('post')){
             $id = $request->input('id');
             $model = Group::find($id);
+            if(!Role::granted('ref_doc_del')){
+                $msg = 'Попытка удаления группы контрагентов '. $model->name;
+                //вызываем event
+                event(new AddEventLogs('access',Auth::id(),$msg));
+                return 'NO';
+            }
             if($model->delete()) {
+                $msg = 'Группа контрагентов '. $model->name .' была удалена!';
+                //вызываем event
+                event(new AddEventLogs('info',Auth::id(),$msg));
                 return 'OK';
             }
             else{
